Modified Worldspace Item Buttons
- It’s messy, but the item pickup buttons are working now.
- Items instantiate buttons based on a prefab when the player is within
range, touching the button adds the item to the player inventory and
deactivates the button
- HOW: had to access the button from a script on an object OTHER THAN
the button itself. No idea why, but the script couldn’t ID a button
component when it was a component on the same GameObject. Could be poor
scripting on my part. onClick.AddListener( function ) works perfectly.