Editor Fixes, Updated blend object and room boundary prefabs for Editor, Scenery Models
Fixed an issue with the room editor not saving the combined mesh to the
correct location.
I updated the room boundary prefab so the lightmap should bake correctly
now. The Wall Section prefab was also updated so the blend surface
components is tagged correctly.
I also included a few scenery models from sketchup to see how they look
with the new level style.