Fixes to Restore Point. Upgrades Screen access point
Tried several times to get this push through, first time the internet
hit the data cap, second the file size for the new level was too big and
I had to remove a room. Final time, there was a merge conflict on the
Restore Point level after I implemented the Upgrade Screen changes for
some reason.

Changes are as follows:
-Created basic movement tutorial level.
-Created Colour Sensor Level 2.
-Reduced file size of new level by removing room.
-Created the Upgrade Trigger Area in the Restore Point level.
-Created GUI instructions in Restore Point level for using Upgrades.
-Made transition to Upgrades Screen scene.
-Made transition back to Restore Point level.
-Duplicated Restore Point in an attempt to fix merge conflict.

NOTE: Upgrade Trigger area knows when an item has been dropped inside
it, but the colliders seem to be incorrect on the data pickup items, as
they clip into the trigger when they are being picked up by the player
due to the scaling effect.